Added anchor handling in menu planner edit mode and some minor fixes

+++ app/controllers/ingredient_controller.rb
@@ -62,6 +62,11 @@ MTMD::FamilyCookBook::App.controllers :ingredient do
       redirect_to url(:ingredient, :index)
     end
 
+    if ingredient.used_in_shopping_lists.any?
+      flash[:error] = "The ingredient is still used in shopping lists! Please remove the references first!"
+      redirect_to url(:ingredient, :index)
+    end
+
     status = @logic_class.destroy
 
     if status == true
